classifier folder structure
