+++ sam_ml/models/classifier/RandomForestClassifier.py
@@ -1,24 +1,24 @@
 from ConfigSpace import Categorical, ConfigurationSpace, Integer, Normal
-from sklearn.ensemble import ExtraTreesClassifier
+from sklearn.ensemble import RandomForestClassifier
 
 from sam_ml.config import get_n_jobs
 
-from .main_classifier import Classifier
+from ..main_classifier import Classifier
 
 
-class ETC(Classifier):
-    """ ExtraTreesClassifier Wrapper class """
+class RFC(Classifier):
+    """ RandomForestClassifier Wrapper class """
 
     def __init__(
         self,
-        model_name: str = "ExtraTreesClassifier",
-        n_jobs: int = get_n_jobs(),
+        model_name: str = "RandomForestClassifier",
+        n_jobs: int = get_n_jobs(), 
         random_state: int = 42,
         **kwargs,
     ):
         """
         @param (important one):
-            n_estimators: Number of trees
+            n_estimators: Number of trees in random forest
             max_depth: Maximum number of levels in tree
             n_jobs: how many cores shall be used (-1 means all)
             random_state: random_state for model
@@ -31,8 +31,8 @@ class ETC(Classifier):
             bootstrap: Method of selecting samples for training each tree
             criterion: function to measure the quality of a split
         """
-        model_type = "ETC"
-        model = ExtraTreesClassifier(
+        model_type = "RFC"
+        model = RandomForestClassifier(
             n_jobs=n_jobs,
             random_state=random_state,
             **kwargs,
@@ -44,7 +44,7 @@ class ETC(Classifier):
             "max_depth": Integer("max_depth", (3, 15), distribution=Normal(5, 3), default=5),
             "min_samples_split": Integer("min_samples_split", (2, 10), default=2),
             "min_samples_leaf": Integer("min_samples_leaf", (1, 4), default=1),
-            "bootstrap": Categorical("bootstrap", [True, False], default=False),
+            "bootstrap": Categorical("bootstrap", [True, False], default=True),
             "criterion": Categorical("criterion", ["gini", "entropy"], default="gini"),
             })
         
@@ -56,7 +56,7 @@ class ETC(Classifier):
             "max_depth": Integer("max_depth", (3, 15), default=5),
             "min_samples_split": Integer("min_samples_split", (2, 10), default=2),
             "min_samples_leaf": Integer("min_samples_leaf", (1, 4), default=1),
-            "bootstrap": Categorical("bootstrap", [True, False], default=False),
+            "bootstrap": Categorical("bootstrap", [True, False], default=True),
             "criterion": Categorical("criterion", ["gini", "entropy"], default="gini"),
             })
         super().__init__(model, model_name, model_type, grid)
