+++ sam_ml/models/classifier/XGBoostClassifier.py
@@ -3,7 +3,7 @@ from xgboost import XGBClassifier
 
 from sam_ml.config import get_n_jobs
 
-from .main_classifier import Classifier
+from ..main_classifier import Classifier
 
 
 class XGBC(Classifier):
